Split components into its own file in platformer example
